# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  availableBalance: number;
  ledgerBalance: number;
  accountNumber: string;
  accountName: string;
  bankName?: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  success: 'success',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface Transaction {
  id: string;
  type: TransactionType;
  title: string;
  /** @nullable */
  description?: string | null;
  amount: number;
  category: string;
  timestamp: string;
  status: TransactionStatus;
  /** @nullable */
  reference?: string | null;
}

export interface TransactionSummary {
  totalInflow: number;
  totalOutflow: number;
  netGrowth: number;
  inflowChange: number;
  outflowChange: number;
}

export interface Network {
  id: string;
  name: string;
  logo: string;
  /** @nullable */
  color?: string | null;
}

export interface AirtimeInput {
  networkId: string;
  phoneNumber: string;
  amount: number;
}

export interface DataPlan {
  id: string;
  title: string;
  validity: string;
  price: number;
}

export interface DataPurchaseInput {
  networkId: string;
  phoneNumber: string;
  planId: string;
}

export interface BillCategory {
  id: string;
  name: string;
  icon: string;
}

export interface BillProvider {
  id: string;
  name: string;
  categoryId: string;
  /** @nullable */
  logo?: string | null;
}

export interface BillPackage {
  id: string;
  name: string;
  amount: number;
  /** @nullable */
  description?: string | null;
}

export interface BillValidateInput {
  providerId: string;
  customerRef: string;
}

export interface BillValidateResult {
  valid: boolean;
  customerName: string;
  /** @nullable */
  address?: string | null;
}

export interface BillPayInput {
  providerId: string;
  customerRef: string;
  /** @nullable */
  packageId?: string | null;
  amount: number;
}

export interface Bank {
  id: string;
  name: string;
}

export interface AccountResolution {
  accountName: string;
  accountNumber: string;
  /** @nullable */
  bankName?: string | null;
}

export type TransferInputType = typeof TransferInputType[keyof typeof TransferInputType];


export const TransferInputType = {
  nex: 'nex',
  bank: 'bank',
} as const;

export interface TransferInput {
  type: TransferInputType;
  /** @nullable */
  recipientId?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  /** @nullable */
  bankCode?: string | null;
  amount: number;
  /** @nullable */
  narration?: string | null;
}

export interface NexUser {
  id: string;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  accountNumber?: string | null;
}

export interface QrDecodeInput {
  data: string;
}

export interface QrDecodeResult {
  type: string;
  amount: number;
  recipientName: string;
  /** @nullable */
  recipientId?: string | null;
  /** @nullable */
  accountNumber?: string | null;
  /** @nullable */
  bankName?: string | null;
}

export interface ScanPayInput {
  qrData: string;
  amount: number;
  /** @nullable */
  narration?: string | null;
}

export interface PurchaseResult {
  success: boolean;
  reference: string;
  message: string;
  /** @nullable */
  transactionId?: string | null;
}

export type GetTransactionsParams = {
limit?: number;
offset?: number;
};

export type GetDataPlansParams = {
network: string;
};

export type GetBillProvidersParams = {
category: string;
};

export type GetBillPackagesParams = {
provider: string;
};

export type ResolveAccountParams = {
accountNumber: string;
bankCode: string;
};

export type SearchUsersParams = {
q: string;
};