# lib/api-zod/src/generated/types/airtimeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface AirtimeInput {
  networkId: string;
  phoneNumber: string;
  amount: number;
}