# lib/api-zod/src/generated/types/billValidateInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface BillValidateInput {
  providerId: string;
  customerRef: string;
}