# lib/api-zod/src/generated/types/billValidateResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface BillValidateResult {
  valid: boolean;
  customerName: string;
  /** @nullable */
  address?: string | null;
}