# lib/api-zod/src/generated/types/dataPlan.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface DataPlan {
  id: string;
  title: string;
  validity: string;
  price: number;
}