# lib/api-zod/src/generated/types/network.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface Network {
  id: string;
  name: string;
  logo: string;
  /** @nullable */
  color?: string | null;
}