# lib/api-zod/src/generated/types/nexUser.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Nex Monie fintech API
 * OpenAPI spec version: 0.1.0
 */

export interface NexUser {
  id: string;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  accountNumber?: string | null;
}